Fix new warnings and suggestions
